+++ layout/src/main/java/com/hortonworks/iotas/layout/design/rule/Rule.java
@@ -21,6 +21,7 @@ package com.hortonworks.iotas.layout.design.rule;
 import com.google.common.collect.ImmutableList;
 import com.hortonworks.iotas.layout.design.rule.action.Action;
 import com.hortonworks.iotas.layout.design.rule.condition.Condition;
+import com.hortonworks.iotas.layout.design.rule.condition.Projection;
 
 import java.io.Serializable;
 import java.util.List;
@@ -35,6 +36,7 @@ public class Rule implements Serializable {
     private String description;
     private String ruleProcessorName;
 
+    private Projection projection;
     private Condition condition;
     private List<Action> actions;
 
@@ -86,12 +88,24 @@ public class Rule implements Serializable {
         this.condition = condition;
     }
 
+    public Projection getProjection() {
+        return projection;
+    }
+
+    public void setProjection(Projection projection) {
+        this.projection = projection;
+    }
+
     public List<Action> getActions() {
         return actions;
     }
 
     public void setActions(List<Action> actions) {
         this.actions = ImmutableList.copyOf(actions);
+    }
+
+    public String getOutputStreamNameForAction(Action action) {
+        return ruleProcessorName + "." + name + "." + id + "." + action.getName();
     }
 
     @Override