+++ webservice/src/main/java/com/hortonworks/iotas/webservice/catalog/dto/DataSourceDto.java
@@ -27,10 +27,12 @@ public class DataSourceDto {
      */
     private String description;
 
+
+
     /**
-     * tag ids of the tag associated with this datasource.
+     * Support on the fly tag creation and make it backward compatible
      */
-    private List<Long> tagIds;
+    private String tags;
 
     /**
      * Time when this entry was created/updated.
@@ -76,7 +78,7 @@ public class DataSourceDto {
         dataSourceId = dataSource.getId();
         dataSourceName = dataSource.getName();
         description = dataSource.getDescription();
-        tagIds = getTagIds(dataSource.getTags());
+        tags = getTagNames(dataSource.getTags());
         timestamp = dataSource.getTimestamp();
         type = dataSource.getType();
         typeConfig = dataSource.getTypeConfig();
@@ -87,6 +89,10 @@ public class DataSourceDto {
             dataFeedType = dataFeed.getType();
         }
     }
+
+    public String getTags() { return tags; }
+
+    public void setTags(String tags) { this.tags = tags; }
 
     public Long getDataSourceId() {
         return dataSourceId;
@@ -110,14 +116,6 @@ public class DataSourceDto {
 
     public void setDescription(String description) {
         this.description = description;
-    }
-
-    public List<Long> getTagIds() {
-        return tagIds;
-    }
-
-    public void setTagIds(List<Long> tagIds) {
-        this.tagIds = tagIds;
     }
 
     public Long getTimestamp() {
@@ -214,7 +212,7 @@ public class DataSourceDto {
                 "dataSourceId=" + dataSourceId +
                 ", dataSourceName='" + dataSourceName + '\'' +
                 ", description='" + description + '\'' +
-                ", tagIds='" + tagIds + '\'' +
+                ", tags='" + tags + '\'' +
                 ", timestamp=" + timestamp +
                 ", type=" + type +
                 ", typeConfig='" + typeConfig + '\'' +
@@ -224,13 +222,14 @@ public class DataSourceDto {
                 '}';
     }
 
-    private List<Long> getTagIds(List<Tag> tags) {
-        List<Long> tagIds = new ArrayList<>();
-        if (tags != null) {
+    private String getTagNames (List<Tag> tags) {
+        StringBuilder tagNames = new StringBuilder();
+        if (tags != null && !tags.isEmpty()) {
             for (Tag tag : tags) {
-                tagIds.add(tag.getId());
+                tagNames.append(tag.getName()).append(",");
             }
+            tagNames.deleteCharAt(tagNames.length() - 1);
         }
-        return tagIds;
+        return tagNames.toString();
     }
 }